+++ supertokens_python/recipe/passwordless/recipe_implementation.py
@@ -31,7 +31,9 @@ from .interfaces import (
     ConsumeCodeRestartFlowErrorResult, CreateCodeOkResult,
     CreateNewCodeForDeviceOkResult,
     CreateNewCodeForDeviceRestartFlowErrorResult,
-    CreateNewCodeForDeviceUserInputCodeAlreadyUsedErrorResult, RecipeInterface,
+    CreateNewCodeForDeviceUserInputCodeAlreadyUsedErrorResult,
+    DeleteUserInfoOkResult, DeleteUserInfoResult,
+    DeleteUserInfoUnknownUserIdErrorResult, RecipeInterface,
     RevokeAllCodesOkResult, RevokeCodeOkResult,
     UpdateUserEmailAlreadyExistsErrorResult, UpdateUserOkResult,
     UpdateUserPhoneNumberAlreadyExistsErrorResult,
@@ -225,6 +227,28 @@ class RecipeImplementation(RecipeInterface):
         if result['status'] == 'EMAIL_ALREADY_EXISTS_ERROR':
             return UpdateUserEmailAlreadyExistsErrorResult()
         return UpdateUserPhoneNumberAlreadyExistsErrorResult()
+
+    async def delete_email_for_user(self, user_id: str, user_context: Dict[str, Any]) -> DeleteUserInfoResult:
+        data = {'userId': user_id, 'email': None}
+        result = await self.querier.send_put_request(NormalisedURLPath('/recipe/user'), data)
+        if result['status'] == 'OK':
+            return DeleteUserInfoOkResult()
+        if result.get("EMAIL_ALREADY_EXISTS_ERROR"):
+            raise Exception("Should never come here")
+        if result.get("PHONE_NUMBER_ALREADY_EXISTS_ERROR"):
+            raise Exception("Should never come here")
+        return DeleteUserInfoUnknownUserIdErrorResult()
+
+    async def delete_phone_number_for_user(self, user_id: str, user_context: Dict[str, Any]) -> DeleteUserInfoResult:
+        data = {'userId': user_id, 'phoneNumber': None}
+        result = await self.querier.send_put_request(NormalisedURLPath('/recipe/user'), data)
+        if result['status'] == 'OK':
+            return DeleteUserInfoOkResult()
+        if result.get("EMAIL_ALREADY_EXISTS_ERROR"):
+            raise Exception("Should never come here")
+        if result.get("PHONE_NUMBER_ALREADY_EXISTS_ERROR"):
+            raise Exception("Should never come here")
+        return DeleteUserInfoUnknownUserIdErrorResult()
 
     async def revoke_all_codes(self,
                                email: Union[str, None], phone_number: Union[str, None], user_context: Dict[str, Any]) -> RevokeAllCodesResult: