+++ supertokens_python/recipe/thirdparty/providers/discord.py
@@ -29,10 +29,11 @@ class Discord(Provider):
                  authorisation_redirect: Union[None, Dict[str, Union[str, Callable[[
                      BaseRequest], str]]]] = None,
                  is_default: bool = False):
-        super().__init__('discord', client_id, is_default)
+        super().__init__('discord', is_default)
         default_scopes = ["email", "identify"]
         if scope is None:
             scope = default_scopes
+        self.client_id = client_id
         self.client_secret = client_secret
         self.base_url = 'https://discord.com'
         self.scopes = list(set(scope))
@@ -80,3 +81,6 @@ class Discord(Provider):
 
     def get_redirect_uri(self, user_context: Dict[str, Any]) -> Union[None, str]:
         return None
+
+    def get_client_id(self, user_context: Dict[str, Any]) -> str:
+        return self.client_id