+++ supertokens_python/recipe/thirdpartypasswordless/asyncio/__init__.py
@@ -14,6 +14,9 @@
 
 from typing import Any, Dict, List, Union
 
+from supertokens_python.recipe.passwordless.interfaces import \
+    DeleteUserInfoResult
+
 from .. import interfaces
 from ..recipe import ThirdPartyPasswordlessRecipe
 from ..types import User
@@ -122,6 +125,18 @@ async def update_passwordless_user(user_id: str,
     if user_context is None:
         user_context = {}
     return await ThirdPartyPasswordlessRecipe.get_instance().recipe_implementation.update_passwordless_user(user_id=user_id, email=email, phone_number=phone_number, user_context=user_context)
+
+
+async def delete_email_for_passwordless_user(user_id: str, user_context: Union[None, Dict[str, Any]] = None) -> DeleteUserInfoResult:
+    if user_context is None:
+        user_context = {}
+    return await ThirdPartyPasswordlessRecipe.get_instance().recipe_implementation.delete_email_for_passwordless_user(user_id=user_id, user_context=user_context)
+
+
+async def delete_phone_number_for_user(user_id: str, user_context: Union[None, Dict[str, Any]] = None) -> DeleteUserInfoResult:
+    if user_context is None:
+        user_context = {}
+    return await ThirdPartyPasswordlessRecipe.get_instance().recipe_implementation.delete_phone_number_for_user(user_id=user_id, user_context=user_context)
 
 
 async def revoke_all_codes(email: Union[str, None] = None,