+++ supertokens_python/recipe/thirdpartypasswordless/syncio/__init__.py
@@ -15,6 +15,8 @@
 from typing import Any, Dict, List, Union
 
 from supertokens_python.async_to_sync_wrapper import sync
+from supertokens_python.recipe.passwordless.interfaces import \
+    DeleteUserInfoResult
 
 from .. import asyncio, interfaces
 from ..types import User
@@ -104,6 +106,14 @@ def update_passwordless_user(user_id: str, email: Union[str, None] = None,
                              phone_number: Union[str, None] = None, user_context: Union[None, Dict[str, Any]] = None) -> interfaces.UpdateUserResult:
     return sync(asyncio.update_passwordless_user(user_id=user_id, email=email,
                                                  phone_number=phone_number, user_context=user_context))
+
+
+def delete_email_for_passwordless_user(user_id: str, user_context: Union[None, Dict[str, Any]] = None) -> DeleteUserInfoResult:
+    return sync(asyncio.delete_email_for_passwordless_user(user_id=user_id, user_context=user_context))
+
+
+def delete_phone_number_for_user(user_id: str, user_context: Union[None, Dict[str, Any]] = None) -> DeleteUserInfoResult:
+    return sync(asyncio.delete_phone_number_for_user(user_id=user_id, user_context=user_context))
 
 
 def revoke_all_codes(email: Union[str, None] = None,