+++ supertokens_python/supertokens.py
@@ -32,7 +32,7 @@ from .recipe.session.cookie_and_header import (
     attach_id_refresh_token_to_cookie_and_header,
     attach_refresh_token_to_cookie, clear_cookies, set_front_token_in_headers)
 from .types import ThirdPartyInfo, User, UsersResponse
-from .utils import (execute_async, get_max_version, get_rid_from_request,
+from .utils import (execute_async, get_rid_from_request, is_version_gte,
                     normalise_http_method, send_non_200_response)
 
 if TYPE_CHECKING:
@@ -283,7 +283,7 @@ class Supertokens:
 
         cdi_version = await querier.get_api_version()
 
-        if get_max_version(cdi_version, "2.10") == cdi_version:
+        if is_version_gte(cdi_version, "2.10"):
             await querier.send_post_request(NormalisedURLPath(USER_DELETE), {
                 "userId": user_id
             })