+++ tests/auth-react/django3x/polls/views.py
@@ -53,8 +53,9 @@ else:
 
 class CustomAuth0Provider(Provider):
     def __init__(self, client_id: str, client_secret: str, domain: str):
-        super().__init__('auth0', client_id, False)
+        super().__init__('auth0', False)
         self.domain = domain
+        self.client_id = client_id
         self.client_secret = client_secret
         self.authorisation_redirect_url = "https://" + self.domain + "/authorize"
         self.access_token_api_url = "https://" + self.domain + "/oauth/token"
@@ -86,6 +87,9 @@ class CustomAuth0Provider(Provider):
 
     def get_redirect_uri(self, user_context: Dict[str, Any]) -> Union[None, str]:
         return None
+
+    def get_client_id(self, user_context: Dict[str, Any]) -> str:
+        return self.client_id
 
 
 async def save_code_text(param: CreateAndSendCustomTextMessageParameters, _: Dict[str, Any]):