+++ tests/utils.py
@@ -20,19 +20,23 @@ from subprocess import DEVNULL, run
 from time import sleep
 from typing import Any, Dict, List
 
-from fastapi.testclient import TestClient
 from requests.models import Response
 from supertokens_python import Supertokens
 from supertokens_python.process_state import ProcessState
 from supertokens_python.recipe.emailpassword import EmailPasswordRecipe
 from supertokens_python.recipe.emailverification import EmailVerificationRecipe
 from supertokens_python.recipe.jwt import JWTRecipe
+from supertokens_python.recipe.passwordless import PasswordlessRecipe
 from supertokens_python.recipe.session import SessionRecipe
 from supertokens_python.recipe.thirdparty import ThirdPartyRecipe
 from supertokens_python.recipe.thirdpartyemailpassword import \
     ThirdPartyEmailPasswordRecipe
+from supertokens_python.recipe.thirdpartypasswordless import \
+    ThirdPartyPasswordlessRecipe
 from supertokens_python.recipe.usermetadata import UserMetadataRecipe
 from yaml import FullLoader, dump, load
+
+from fastapi.testclient import TestClient
 
 INSTALLATION_PATH = environ['SUPERTOKENS_PATH']
 SUPERTOKENS_PROCESS_DIR = INSTALLATION_PATH + '/.started'
@@ -176,8 +180,11 @@ def reset():
     EmailPasswordRecipe.reset()
     EmailVerificationRecipe.reset()
     ThirdPartyRecipe.reset()
+    PasswordlessRecipe.reset()
     JWTRecipe.reset()
     UserMetadataRecipe.reset()
+    ThirdPartyPasswordlessRecipe.reset()
+    PasswordlessRecipe.reset()
 
 
 def get_cookie_from_response(response: Response, cookie_name: str):