+++ CHANGELOG.md
@@ -5,7 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+- Updates Project Setup, Modifying Code and Testing sections in the contributing guide
 - Fixed async execution of `send_telemetry` in init and `call_get_handshake_info` in session recipe implementation.
+
+## [0.7.0] - 2022-04-28
+- Changes third party provider type to get client ID dynamically so that it can be changed based on user context.
+
+## [0.6.7] - 2022-04-23
+- Adds delete email (`delete_email_for_user`) and phone number (`delete_phone_number_for_user`) functions for passwordless and thirdpartypasswordless recipe
+- Adds check for user type in update passwordless info in thirdpartypasswordless recipe
+
+## [0.6.6] - 2022-04-22
+- Fixes issue in user metadata recipe where as are exposing async functions in the syncio file.
 
 ## [0.6.5] - 2022-04-18
 - Upgrade and freeze pyright version